--- app/models/Comment.rb
@@ -1,4 +1,0 @@
-class Comment < ApplicationRecord
-  belongs_to :User
-  belongs_to :Post
-end
