--- app/models/Like.rb
@@ -1,4 +1,0 @@
-class Like < ApplicationRecord
-  belongs_to :User
-  belongs_to :Post
-end
