+++ app/models/post.rb
@@ -1,3 +1,3 @@
 class Post < ApplicationRecord
-  belongs_to :User
+  belongs_to :author
 end
