--- db/migrate/20220610093428_create_likes.rb
@@ -1,10 +1,0 @@
-class CreateLikes < ActiveRecord::Migration[7.0]
-  def change
-    create_table :likes do |t|
-      t.references :author, null: false, foreign_key: {to_table: 'users'}
-      t.references :posts, null: false, foreign_key: {to_table: 'posts'}
-
-      t.timestamps
-    end
-  end
-end
