+++ db/migrate/20220613150120_create_users.rb
