+++ db/migrate/20220613150447_create_posts.rb
@@ -1,7 +1,7 @@
 class CreatePosts < ActiveRecord::Migration[7.0]
   def change
     create_table :posts do |t|
-      t.references :author, null: false, foreign_key: {to_table: 'users'}
+      t.references :author, null: false, foreign_key:{to_table: 'users'}
       t.string :title
       t.text :text
       t.integer :comments_counter
