+++ db/migrate/20220613150753_create_comments.rb
@@ -1,8 +1,8 @@
 class CreateComments < ActiveRecord::Migration[7.0]
   def change
     create_table :comments do |t|
-      t.references :author, null: false, foreign_key: {to_table: 'users'}
-      t.references :posts, null: false, foreign_key: {to_table: 'posts'}
+      t.references :author, null: false, foreign_key:{to_table: 'users'}
+      t.references :posts, null: false, foreign_key:{to_table: 'posts'}
       t.string :text
 
       t.timestamps
